// src/types.ts
export interface EditableContent {
  // Image URLs
  heroBg: string;
  storefront: string;
  makima: string;
  aki: string;
  kobeni: string;
  makimaDenji: string;
  angelDevil: string;

  // Texts
  heroTitle: string;
  heroDesc: string;
  section1Title: string;
  makimaCardTitle: string;
  makimaCardDesc: string;
  akiCardTitle: string;
  akiCardDesc: string;
  kobeniCardTitle: string;
  kobeniCardDesc: string;
  section2Title: string;
  section2Percent: string;
  section2Quote: string;
  trailerTitle: string;
  trailerSub: string;
  trailerDesc: string;
  trailerYoutubeUrl: string;
}